Move shutdown handlers to proper source file
