+++ start.c
@@ -2,7 +2,6 @@
 
 #include <curl/curl.h>
 #include <omp.h>
-#include <signal.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <sys/stat.h>
@@ -16,10 +15,6 @@
 #include "logger.h"
 #include "shutdown.h"
 #include "types.h"
-
-#if _IS_WINDOWS
-#include <windows.h>
-#endif
 
 #define UNSET_FRAME_RECORD 9999
 
@@ -41,52 +36,6 @@ void setLocalRecord(int frames) {
 		return;
 	}
 	current_frame_record = frames;
-}
-
-int numTimesExitRequest = 0;
-#define NUM_TIMES_EXITED_BEFORE_HARD_QUIT 3
-
-void countAndSetShutdown(bool isSignal) {
-	// On Windows, it is undefined behavior trying to use stdio.h functions in a signal handler (which this is called from).
-	// So for now, these messages are stifled on Windows. This may be revisited at a later point.
-	if (++numTimesExitRequest >= NUM_TIMES_EXITED_BEFORE_HARD_QUIT) {
-		if (!_IS_WINDOWS || !isSignal) {
-			printf("\nExit reqested %d times; shutting down now.\n", NUM_TIMES_EXITED_BEFORE_HARD_QUIT);
-		}
-		exit(1);
-	} else {
-		requestShutdown();
-		if (!_IS_WINDOWS || !isSignal) {
-			printf("\nExit requested, finishing up work. Should shutdown soon (CTRL-C %d times total to force exit)\n", NUM_TIMES_EXITED_BEFORE_HARD_QUIT);
-		}
-	}
-}
-
-void handleTermSignal(int signal) {
-	countAndSetShutdown(true);
-}
-
-#if _IS_WINDOWS
-BOOL WINAPI windowsCtrlCHandler(DWORD fdwCtrlType) {
-	switch (fdwCtrlType) {
-	case CTRL_C_EVENT: ABSL_FALLTHROUGH_INTENDED;
-	case CTRL_CLOSE_EVENT:
-		countAndSetShutdown(false);
-		return TRUE;
-	default:
-		return FALSE;
-	}
-}
-#endif
-
-void setSignalHandlers() {
-	signal(SIGTERM, handleTermSignal);
-	signal(SIGINT, handleTermSignal);
-#if _IS_WINDOWS
-	if (!SetConsoleCtrlHandler(windowsCtrlCHandler, TRUE)) {
-		printf("Unable to set CTRL-C handler. CTRL-C may cause unclean shutdown.\n");
-	}
-#endif
 }
 
 void printAsciiGreeting()
